Move Date Transformers to General Transformer Class
