+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/Transformers.java
@@ -2,8 +2,13 @@ package gov.va.api.health.dataquery.service.controller;
 
 import static org.springframework.util.CollectionUtils.isEmpty;
 
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
 import java.util.Collection;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -71,6 +76,40 @@ public final class Transformers {
       return null;
     }
     return extract.apply(object);
+  }
+
+  /** Return null if the date is null, otherwise return an ISO-8601 date. */
+  public static String asDateString(Optional<LocalDate> maybeDateTime) {
+    if (maybeDateTime == null) {
+      return null;
+    }
+    return asDateString(maybeDateTime.orElse(null));
+  }
+
+  /** Return null if the date is null, otherwise return an ISO-8601 date. */
+  public static String asDateString(LocalDate maybeDateTime) {
+    if (maybeDateTime == null) {
+      return null;
+    }
+    return DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.US)
+            .withZone(ZoneOffset.UTC)
+            .format(maybeDateTime);
+  }
+
+  /** Return null if the date is null, otherwise return an ISO-8601 date time. */
+  public static String asDateTimeString(Optional<Instant> maybeDateTime) {
+    if (maybeDateTime == null) {
+      return null;
+    }
+    return asDateTimeString(maybeDateTime.orElse(null));
+  }
+
+  /** Return null if the date is null, otherwise return an ISO-8601 date time. */
+  public static String asDateTimeString(Instant maybeDateTime) {
+    if (maybeDateTime == null) {
+      return null;
+    }
+    return maybeDateTime.toString();
   }
 
   /** Return true if the value is a blank string, or any other object that is null. */
