+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/immunization/Dstu2ImmunizationTransformer.java
@@ -1,7 +1,7 @@
 package gov.va.api.health.dataquery.service.controller.immunization;
 
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateTimeString;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReference;
+import static gov.va.api.health.dataquery.service.controller.Transformers.asDateTimeString;
 
 import gov.va.api.health.argonaut.api.resources.Immunization;
 import gov.va.api.health.dataquery.service.controller.datamart.DatamartReference;
