+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/procedure/Dstu2ProcedureTransformer.java
@@ -1,8 +1,8 @@
 package gov.va.api.health.dataquery.service.controller.procedure;
 
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asCodeableConceptWrapping;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateTimeString;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReference;
+import static gov.va.api.health.dataquery.service.controller.Transformers.asDateTimeString;
 
 import gov.va.api.health.argonaut.api.resources.Procedure;
 import gov.va.api.health.argonaut.api.resources.Procedure.Status;
