+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/Dstu2TransformersTest.java
@@ -3,8 +3,6 @@ package gov.va.api.health.dataquery.service.controller;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asCodeableConceptWrapping;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asCoding;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDatamartReference;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateString;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateTimeString;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReference;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReferenceId;
 import static gov.va.api.health.dataquery.service.controller.Transformers.convert;
@@ -17,8 +15,6 @@ import gov.va.api.health.dataquery.service.controller.datamart.DatamartReference
 import gov.va.api.health.dstu2.api.datatypes.CodeableConcept;
 import gov.va.api.health.dstu2.api.datatypes.Coding;
 import gov.va.api.health.dstu2.api.elements.Reference;
-import java.time.Instant;
-import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
@@ -90,60 +86,6 @@ public class Dstu2TransformersTest {
     assertThat(dataRef.type().get()).isEqualTo("Patient");
     assertThat(dataRef.reference().get()).isEqualTo("123V456");
     assertThat(dataRef.display().get()).isEqualTo("WAT");
-  }
-
-  @Test
-  public void asDateStringReturnsNullWhenInstantIsNull() {
-    assertThat(asDateString((LocalDate) null)).isNull();
-  }
-
-  @Test
-  public void asDateStringReturnsNullWhenOptionalInstantIsEmpty() {
-    assertThat(asDateString(Optional.empty())).isNull();
-  }
-
-  @Test
-  public void asDateStringReturnsNullWhenOptionalInstantIsNull() {
-    assertThat(asDateString((Optional<LocalDate>) null)).isNull();
-  }
-
-  @Test
-  public void asDateStringReturnsStringWhenInstantIsNotNull() {
-    LocalDate time = LocalDate.parse("2005-01-21");
-    assertThat(asDateString(time)).isEqualTo("2005-01-21");
-  }
-
-  @Test
-  public void asDateStringReturnsStringWhenOptionalInstantIsNotNull() {
-    LocalDate time = LocalDate.parse("2005-01-21");
-    assertThat(asDateString(Optional.of(time))).isEqualTo("2005-01-21");
-  }
-
-  @Test
-  public void asDateTimeStringReturnsNullWhenInstantIsNull() {
-    assertThat(asDateTimeString((Instant) null)).isNull();
-  }
-
-  @Test
-  public void asDateTimeStringReturnsNullWhenOptionalInstantIsEmpty() {
-    assertThat(asDateTimeString(Optional.empty())).isNull();
-  }
-
-  @Test
-  public void asDateTimeStringReturnsNullWhenOptionalInstantIsNull() {
-    assertThat(asDateTimeString((Optional<Instant>) null)).isNull();
-  }
-
-  @Test
-  public void asDateTimeStringReturnsStringWhenInstantIsNotNull() {
-    Instant time = Instant.parse("2005-01-21T07:57:00.000Z");
-    assertThat(asDateTimeString(time)).isEqualTo("2005-01-21T07:57:00Z");
-  }
-
-  @Test
-  public void asDateTimeStringReturnsStringWhenOptionalInstantIsPresent() {
-    Instant time = Instant.parse("2005-01-21T07:57:00.000Z");
-    assertThat(asDateTimeString(Optional.of(time))).isEqualTo("2005-01-21T07:57:00Z");
   }
 
   @Test
